refactor oauth  integration settings tab components
