+++ apps/studio/components/interfaces/Integrations/Integration/MarketplaceIntegrationSettingsTab.tsx
@@ -1,12 +1,17 @@
-import dayjs from 'dayjs'
-import { Settings, Trash2, TriangleAlert } from 'lucide-react'
-import { useState, type ReactNode } from 'react'
+import { useState } from 'react'
 import { toast } from 'sonner'
-import { Badge, Button } from 'ui'
 import { Admonition, GenericSkeletonLoader } from 'ui-patterns'
 import ConfirmationModal from 'ui-patterns/Dialogs/ConfirmationModal'
 
-import { defaultDisabledSmtpFormValues } from '@/components/interfaces/Auth/SmtpForm/SmtpForm.constants'
+import { ResourceGroupSection } from './ConnectedResourceGroupSection'
+import { type ApiKeyResource, type ResourceGroup } from './MarketplaceIntegrationSettingsTab.types'
+import {
+  getGroupContent,
+  getItemIdentifier,
+  getItemMeta,
+  KIND_ORDER,
+} from './MarketplaceIntegrationSettingsTab.utils'
+import { useConnectedResourceMutations } from './useConnectedResourceMutations'
 import { ConstrainedIntegrationTabScaffold } from '@/components/interfaces/Integrations/ConstrainedIntegrationTabScaffold'
 import {
   getConnectedResources,
@@ -14,263 +19,12 @@ import {
   getExpectedResourceKinds,
   useProjectOAuthIntegrationData,
   type ConnectedResource,
-  type ConnectedResourceUsage,
 } from '@/components/interfaces/Integrations/Landing/Landing.utils'
 import { useIntegrationDetail } from '@/components/interfaces/Integrations/Landing/useIntegrationDetail'
 import { RevokeAppModal } from '@/components/interfaces/Organization/OAuthApps/RevokeAppModal'
 import { TextConfirmModal } from '@/components/ui/TextConfirmModalWrapper'
-import { useAPIKeyDeleteMutation } from '@/data/api-keys/api-key-delete-mutation'
-import { useAuthConfigUpdateMutation } from '@/data/auth/auth-config-update-mutation'
-import { useAuthorizedAppRevokeMutation } from '@/data/oauth/authorized-app-revoke-mutation'
 import type { AuthorizedApp } from '@/data/oauth/authorized-apps-query'
-import { useSecretsDeleteMutation } from '@/data/secrets/secrets-delete-mutation'
 import { useSelectedOrganizationQuery } from '@/hooks/misc/useSelectedOrganization'
-
-type ResourceKind = ConnectedResource['kind']
-type ApiKeyResource = Extract<ConnectedResource, { kind: 'api_key' }>
-
-type ManageAction = { label: string; href: string }
-
-/** A single removable entry within a section (one OAuth app, one API key, one secret, etc.). */
-type ResourceItem = { resource: ConnectedResource; identifier: string; meta?: string }
-
-/** A group of same-kind resources rendered as one section (e.g. all secret API keys together). */
-type ResourceGroup = {
-  kind: ResourceKind
-  title: string
-  badge?: string
-  description: ReactNode
-  note: string
-  /** Impact copy shown in the zero state when the resource is expected but absent. */
-  missingNote: string
-  manageAction?: ManageAction
-  items: ResourceItem[]
-  /** True when the integration expects this resource but none is currently connected. */
-  missing?: boolean
-}
-
-const KIND_ORDER: ResourceKind[] = ['oauth_app', 'api_key', 'edge_function_secret', 'smtp']
-
-const formatDate = (value?: string | null) =>
-  value ? dayjs(value).format('MMM D, YYYY') : undefined
-
-/** The monospace identifier shown for an individual resource. */
-const getItemIdentifier = (resource: ConnectedResource): string => {
-  switch (resource.kind) {
-    case 'oauth_app':
-      return resource.app.name
-    case 'api_key':
-      return resource.apiKey.name
-    case 'edge_function_secret':
-      return resource.secret.name
-    case 'smtp':
-      return 'smtp.resend.com'
-  }
-}
-
-/** Secondary metadata (timestamp) shown beneath an individual resource. */
-const getItemMeta = (resource: ConnectedResource): string | undefined => {
-  switch (resource.kind) {
-    case 'oauth_app': {
-      const date = formatDate(resource.app.authorized_at)
-      return date ? `Authorized ${date}` : undefined
-    }
-    case 'api_key': {
-      const date = formatDate(resource.apiKey.inserted_at)
-      return date ? `Created ${date}` : undefined
-    }
-    case 'edge_function_secret': {
-      const date = formatDate(resource.secret.updated_at)
-      return date ? `Updated ${date}` : undefined
-    }
-    case 'smtp':
-      return undefined
-  }
-}
-
-/**
- * Section-level, integration-aware copy for a kind of resource. The settings page reads like a
- * guide that explains what each resource does and what removing it affects.
- */
-const getGroupContent = ({
-  kind,
-  count,
-  integrationName,
-  orgSlug,
-  projectRef,
-  usage,
-}: {
-  kind: ResourceKind
-  count: number
-  integrationName: string
-  orgSlug?: string
-  projectRef?: string
-  /** Integration-specific copy that overrides the generic, kind-level description and note. */
-  usage?: ConnectedResourceUsage
-}): Omit<ResourceGroup, 'kind' | 'items' | 'missing'> => {
-  const name = <span className="text-foreground">{integrationName}</span>
-  const plural = count > 1
-
-  switch (kind) {
-    case 'oauth_app':
-      return {
-        title: 'OAuth application',
-        badge: 'Connected',
-        description: usage?.description ?? (
-          <>
-            Grants {name} access to your organization and its projects through a scoped OAuth grant.
-          </>
-        ),
-        note:
-          usage?.removalWarning ??
-          'Removing this OAuth app will remove it for all projects and members of your organization.',
-        missingNote:
-          usage?.noteWhenAbsent ??
-          `No OAuth app is connected for ${integrationName}. It does not have access to your organization or its projects.`,
-        manageAction: orgSlug
-          ? { label: 'Manage access', href: `/org/${orgSlug}/apps` }
-          : undefined,
-      }
-    case 'api_key':
-      return {
-        title: plural ? 'Secret API keys' : 'Secret API key',
-        description:
-          usage?.description ??
-          (plural ? (
-            <>Secret API keys that {name} uses to authenticate to your project&apos;s API.</>
-          ) : (
-            <>A secret API key that {name} uses to authenticate to your project&apos;s API.</>
-          )),
-        note:
-          usage?.removalWarning ??
-          `Removing ${plural ? 'a key' : 'this key'} takes effect immediately and can interrupt the integration.`,
-        missingNote:
-          usage?.noteWhenAbsent ??
-          `No secret API key is connected for ${integrationName}. It cannot authenticate to your project's API.`,
-      }
-    case 'edge_function_secret':
-      return {
-        title: plural ? 'Edge Function secrets' : 'Edge Function secret',
-        description:
-          usage?.description ??
-          (plural ? (
-            <>
-              Secrets synced by {name} and exposed to your project&apos;s Edge Functions at runtime.
-            </>
-          ) : (
-            <>
-              A secret synced by {name} and exposed to your project&apos;s Edge Functions at
-              runtime.
-            </>
-          )),
-        note:
-          usage?.removalWarning ??
-          `Removing ${plural ? 'a secret' : 'this secret'} takes effect immediately and can interrupt the integration.`,
-        missingNote:
-          usage?.noteWhenAbsent ??
-          `No Edge Function secret from ${integrationName} was detected. Functions that rely on it may fail at runtime.`,
-      }
-    case 'smtp':
-      return {
-        title: 'SMTP settings',
-        description: usage?.description ?? (
-          <>A custom SMTP relay so your project sends emails through {name}.</>
-        ),
-        note:
-          usage?.removalWarning ??
-          'Removing this relay reverts your project to the default email service. Auth emails may be rate-limited until SMTP is configured again.',
-        missingNote:
-          usage?.noteWhenAbsent ??
-          `SMTP settings for ${integrationName} were not detected. Authentication emails may not be sent through the integration's SMTP service.`,
-        manageAction: projectRef
-          ? { label: 'Manage settings', href: `/project/${projectRef}/auth/smtp` }
-          : undefined,
-      }
-  }
-}
-
-const ResourceGroupSection = ({
-  group,
-  onRemove,
-}: {
-  group: ResourceGroup
-  onRemove: (resource: ConnectedResource) => void
-}) => {
-  return (
-    <section className="flex flex-col gap-y-4 border-b py-8 first:pt-0 last:border-b-0">
-      <div className="flex flex-col gap-y-2">
-        <div className="flex items-center gap-x-2">
-          <h3 className="text-base text-foreground">{group.title}</h3>
-          {group.missing ? (
-            <Badge variant="warning" className="gap-x-1.5">
-              <TriangleAlert size={12} strokeWidth={1.5} />
-              Not connected
-            </Badge>
-          ) : (
-            group.badge && (
-              <Badge variant="success" className="gap-x-1.5">
-                <span className="h-1.5 w-1.5 rounded-full bg-brand" />
-                {group.badge}
-              </Badge>
-            )
-          )}
-        </div>
-        <p className="text-sm text-foreground-light max-w-2xl">{group.description}</p>
-      </div>
-
-      <Admonition
-        type={group.missing ? 'warning' : 'default'}
-        className="m-0 max-w-2xl"
-        title={group.missing ? 'This resource is missing' : undefined}
-      >
-        {group.missing ? group.missingNote : group.note}
-      </Admonition>
-
-      {group.missing ? (
-        group.manageAction && (
-          <div className="max-w-2xl">
-            <Button asChild type="default" icon={<Settings />}>
-              <a href={group.manageAction.href}>{group.manageAction.label}</a>
-            </Button>
-          </div>
-        )
-      ) : (
-        <div className="max-w-2xl divide-y rounded-md border bg-surface-100">
-          {group.items.map((item) => (
-            <div
-              key={item.resource.key}
-              className="flex items-center justify-between gap-x-4 px-4 py-3"
-            >
-              <div className="flex min-w-0 flex-col">
-                <code
-                  className="truncate font-mono text-sm text-foreground"
-                  title={item.identifier}
-                >
-                  {item.identifier}
-                </code>
-                {item.meta && <span className="text-xs text-foreground-lighter">{item.meta}</span>}
-              </div>
-              <div className="flex shrink-0 items-center gap-x-2">
-                {group.manageAction && (
-                  <Button asChild type="default" icon={<Settings />}>
-                    <a href={group.manageAction.href}>{group.manageAction.label}</a>
-                  </Button>
-                )}
-                <Button
-                  type="default"
-                  icon={<Trash2 className="text-foreground-light" />}
-                  onClick={() => onRemove(item.resource)}
-                >
-                  Remove
-                </Button>
-              </div>
-            </div>
-          ))}
-        </div>
-      )}
-    </section>
-  )
-}
 
 export const MarketplaceIntegrationSettingsTab = () => {
   const { ref, integration } = useIntegrationDetail()
@@ -284,22 +38,12 @@ export const MarketplaceIntegrationSettingsTab = () => {
   const [appToRevoke, setAppToRevoke] = useState<AuthorizedApp | undefined>()
   const [apiKeyToDelete, setApiKeyToDelete] = useState<ApiKeyResource | undefined>()
 
-  const onMutationSuccess = () => {
-    toast.success('Successfully removed the connected resource')
-  }
-
-  const { mutateAsync: revokeAuthorizedApp, isPending: isRevokingApp } =
-    useAuthorizedAppRevokeMutation({ onSuccess: onMutationSuccess })
-  const { mutateAsync: deleteAPIKey, isPending: isDeletingApiKey } = useAPIKeyDeleteMutation({
-    onSuccess: onMutationSuccess,
-  })
-  const { mutateAsync: deleteSecrets, isPending: isDeletingSecret } = useSecretsDeleteMutation({
-    onSuccess: onMutationSuccess,
-  })
-  const { mutateAsync: updateAuthConfig, isPending: isUpdatingAuthConfig } =
-    useAuthConfigUpdateMutation({ onSuccess: onMutationSuccess })
-
-  const isRemoving = isRevokingApp || isDeletingApiKey || isDeletingSecret || isUpdatingAuthConfig
+  const { removeResource, deleteApiKey, isRemoving, isDeletingApiKey } =
+    useConnectedResourceMutations({
+      projectRef: ref,
+      orgSlug: organization?.slug,
+      onSuccess: () => toast.success('Successfully removed the connected resource'),
+    })
 
   const integrationName = integration?.name ?? 'this integration'
   const resources =
@@ -354,23 +98,6 @@ export const MarketplaceIntegrationSettingsTab = () => {
   const hasOtherResources = resources.some((r) => r.kind !== 'oauth_app')
   const showOrphanedResourcesWarning = isOAuthAppMissing && hasOtherResources
 
-  const removeResource = async (resource: ConnectedResource) => {
-    switch (resource.kind) {
-      case 'oauth_app':
-        if (!organization?.slug) throw new Error('Organization is required')
-        return revokeAuthorizedApp({ orgSlug: organization.slug, id: resource.app.id })
-      case 'api_key':
-        if (!ref) throw new Error('Project is required')
-        return deleteAPIKey({ projectRef: ref, id: resource.apiKey.id! })
-      case 'edge_function_secret':
-        if (!ref) throw new Error('Project is required')
-        return deleteSecrets({ projectRef: ref, secrets: [resource.secret.name] })
-      case 'smtp':
-        if (!ref) throw new Error('Project is required')
-        return updateAuthConfig({ projectRef: ref, config: defaultDisabledSmtpFormValues })
-    }
-  }
-
   const onSelectRemove = (resource: ConnectedResource) => {
     // Reuse the same dedicated modals used elsewhere for OAuth apps and API keys.
     if (resource.kind === 'oauth_app') setAppToRevoke(resource.app)
@@ -393,9 +120,9 @@ export const MarketplaceIntegrationSettingsTab = () => {
   }
 
   const onConfirmDeleteApiKey = async () => {
-    if (!apiKeyToDelete || !ref) return
+    if (!apiKeyToDelete) return
     try {
-      await deleteAPIKey({ projectRef: ref, id: apiKeyToDelete.apiKey.id! })
+      await deleteApiKey(apiKeyToDelete.apiKey.id!)
       setApiKeyToDelete(undefined)
     } catch {
       // Error toast is handled by the mutation's default onError.
@@ -441,7 +168,7 @@ export const MarketplaceIntegrationSettingsTab = () => {
                   <Admonition
                     type="warning"
                     title="OAuth application is missing"
-                    description={`No OAuth app is connected for ${integrationName}, but other resources associated with it are still present on your project. ${integrationName} may not work correctly without an OAuth app. Reconnect it, or remove the remaining resources below.`}
+                    description={`No OAuth app is connected for ${integrationName}, but other resources associated with it are still present on your project. ${integrationName} may not work correctly without an OAuth app. Reconnect it, or to fully disconnect the integration, remove the remaining resources below.`}
                     className="mb-8 mt-0"
                   />
                 )}
